Builder: the private methods are no more bound
Assuming private method are always called in a public method, `this` is now transmitted into private methods thanks to `call`. Example
`this.#privMethod(arg1, arg2)`
becomes
`this.#privMethod.call(this, arg1, arg2)`
which is transformed into
`__privMethod.call(this, arg1, arg2)`

The only caveat is when a public method returns the private method without calling it. In this way, the code MUST explicitly use `this.#privMethod.bind(this)`